CAL-416 refactor: improve graceperiod module
Before refactor the gracemodule, the workflow for graceperiod was really
bad:

1 - I call a rest service that return me an ID of a grace task

2 - using gracePeriodService.grace I construct a promise that can't
fail! This promise resolve either if the time of the graceperiod is over
(regarding the grace-period was a fail or not !) either if the user
press cancel. The resolved object told me if it was resolved because of
a cancel (see point 4) or timeout (see point 5). This object also
provide me method to said if there were an error or a success. 3 -
because the promise returned by the gracePeriodService.grace does not
allow to really monitor the success or not of my grace-period, I have to
use gracePeriodLiveNotification.registerListeners that take two
callback (onError, onDone). Why isn't it a promise ????

4- (if the promise returned by action 2 succeed because of a cancel) in
this case I have to make a rest call that could fail to stop the
graceperiod. I have to check myself if this call fail or success and
call myself a method on the resolved object by the
gracePeriodService.grace's promise to notify the user about the fail or
the success of the canceling.

5 - if the promise return by action 2 resolve and say the grace-period
was not canceled, I still need to check that the onError callback given
at step 3 wasn't call to check if the grace-period success or not. Also
to be totally sure that the grace-period is over, I should check that
the onDone callback has been called.

Now the workflow is really more easy and gracePeriodLiveNotification
should be used anymore (it's used by gracePeriodService.grace). Also the
factory notifyOfGracedRequest has been removed and replaced by a method
askUserForCancel inside the gracePeriodService.

Here is the new workflow when using the gracePeriodService:

1 - I call a rest service that return me an ID of a grace task

2 - I call gracePeriodService.grace with the id and with text to display
to the user during the graceperiod, for the button to cancel at the end
of the graceperiod and in case of failure during the graceperiod or
during the cancelling of the graceperiod.

3 - gracePeriodService.grace will return a promise that :

  * resolve if there were no cancel and the graceperiod success with
    {cancelled: false}
  * resolve if there were a cancel tentative that failed with {cancelled:
    true, cancelFailed: true}
  * fail if there were a successful cancel tentative with {cancelled:
    true, cancelFailed: false}
  * fail if there were an error at the task execution at the end of the
    grace period with {cancelled: false}