CAL-405 make calendarCurrentView save wanted range
Before calendarCurrentView wasn't saving the wanted range (ie a month)
but the actual range visible arround a wanted range which is the same as
a the wanted range for all view except for month. When a month is
displayed there are most of the some day of the month before and after
in order to display each week in the month fully.

The fact that we saved the actual range and not the wanted range
provocated a bug for month view, when restoring the view the code
displayed the month before !